rename slash
